+++ components/Sidebar/Sidebar.tsx
@@ -1,7 +1,7 @@
 import {IconFolderPlus, IconMistOff, IconPlus} from "@tabler/icons-react"
 import {ReactNode} from "react"
 import {useTranslation} from "react-i18next"
-import {CloseSidebarButton, OpenSidebarButton} from "./components/OpenCloseButton"
+import {CloseSidebarButton, OpenSidebarButton} from "./components/OpenCloseSidebarButton"
 import Search from "../Search"
